made chorus and flanger Stereo

+++ CSD2c/Effects/Chorus_Effect/utilities.h
@@ -2,18 +2,18 @@
 
 namespace Util
 {
-    inline float mapInRange(float factor, float inputMin, float inputMax, float outputMin, float outputMax)
+    inline double mapInRange(double factor, double inputMin, double inputMax, double outputMin, double outputMax)
     {
-        float inputDelta = inputMax - inputMin;
-        float outputDelta = outputMax - outputMin;
-        float valueScaled = (factor-inputMin)/inputDelta;
-        float output = valueScaled * outputDelta + outputMin;
+        double inputDelta = inputMax - inputMin;
+        double outputDelta = outputMax - outputMin;
+        double valueScaled = (factor-inputMin)/inputDelta;
+        double output = valueScaled * outputDelta + outputMin;
         return output;
     }
 
-    inline float linearMap(float factor, float min, float max)
+    inline double linearMap(double factor, double min, double max)
     {
-        float output = mapInRange(factor, 0.0, 1.0, min, max);
+        double output = mapInRange(factor, 0.0, 1.0, min, max);
         return output;
     }
